move exceptions out of package
